recreate the print_array.c and print_list.c files
