+++ 0-bubble_sort.c
@@ -13,7 +13,7 @@ void bubble_sort(int *array, size_t size)
 	size_t i, j;
 	int swapped;
 
-	if (!array || size <= 1)
+	if (!array || size < 2)
 		return;
 
 	for (i = 0; i < size - 1; i++)
